+++ ds3-cli-helpers/src/main/java/com/spectralogic/ds3cli/helpers/Util.java
@@ -1,6 +1,6 @@
 /*
  * ******************************************************************************
- *   Copyright 2014-2015 Spectra Logic Corporation. All Rights Reserved.
+ *   Copyright 2014-2016 Spectra Logic Corporation. All Rights Reserved.
  *   Licensed under the Apache License, Version 2.0 (the "License"). You may not use
  *   this file except in compliance with the License. A copy of the License is located at
  *
@@ -13,7 +13,7 @@
  * ****************************************************************************
  */
 
-package com.spectralogic.ds3cli.integration;
+package com.spectralogic.ds3cli.helpers;
 
 import com.google.common.collect.ImmutableList;
 import com.spectralogic.ds3cli.*;
