+++ ds3-cli-integration/src/test/java/com/spectralogic/ds3cli/integration/FeatureIntegration_Test.java
@@ -20,10 +20,12 @@ import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Lists;
 import com.spectralogic.ds3cli.Arguments;
 import com.spectralogic.ds3cli.CommandResponse;
+import com.spectralogic.ds3cli.helpers.JsonMapper;
 import com.spectralogic.ds3cli.helpers.TempStorageIds;
 import com.spectralogic.ds3cli.helpers.TempStorageUtil;
-import com.spectralogic.ds3cli.integration.models.HeadObject;
-import com.spectralogic.ds3cli.integration.models.JobResponse;
+import com.spectralogic.ds3cli.helpers.Util;
+import com.spectralogic.ds3cli.helpers.models.HeadObject;
+import com.spectralogic.ds3cli.helpers.models.JobResponse;
 import com.spectralogic.ds3cli.util.FileUtils;
 import com.spectralogic.ds3cli.util.SterilizeString;
 import com.spectralogic.ds3client.Ds3Client;
